+++ src/libtrx/include/libtrx/game/clock.h
@@ -2,3 +2,4 @@
 
 #include "clock/common.h"
 #include "clock/timer.h"
+#include "clock/turbo.h"