+++ src/tr2/game/console/setup.c
@@ -17,13 +17,16 @@
 #include <libtrx/game/console/cmd/save_game.h>
 #include <libtrx/game/console/cmd/set_health.h>
 #include <libtrx/game/console/cmd/sfx.h>
+#include <libtrx/game/console/cmd/speed.h>
 #include <libtrx/game/console/cmd/teleport.h>
 
 #include <stddef.h>
 
 CONSOLE_COMMAND *g_ConsoleCommands[] = {
+    // clang-format off
     &g_Console_Cmd_Teleport,
     &g_Console_Cmd_Fly,
+    &g_Console_Cmd_Speed,
     &g_Console_Cmd_FlipMap,
     &g_Console_Cmd_Kill,
     &g_Console_Cmd_EndLevel,
@@ -40,5 +43,6 @@ CONSOLE_COMMAND *g_ConsoleCommands[] = {
     &g_Console_Cmd_SetHealth,
     &g_Console_Cmd_GiveItem,
     &g_Console_Cmd_SFX,
+    // clang-format on
     NULL,
 };