+++ src/tr2/game/stats.h
@@ -1,0 +1,4 @@
+#pragma once
+
+void Stats_StartTimer(void);
+void Stats_UpdateTimer(void);