+++ docs/tr2/CHANGELOG.md
@@ -2,6 +2,7 @@
 - added a fly cheat key (#1642)
 - added an items cheat key (#1641)
 - added a level skip cheat key (#1640)
+- added a turbo cheat (#1639)
 - added the ability to skip end credits with the action and escape keys (#1800)
 - added the ability to skip FMVs with the action key (#1650)
 - added the ability to hold forward/back to move through menus more quickly (#1644)